Assigning quality to the node is updated. PReviously quality was updated whenever we solver is called.
Now when nodes are created we set their quality to their parents.
After a solver call we update the quality only the objective value is
larger (for minimization problem).

Solver objective values may result qualities lower than the parent quality
due to lack of conic approximation cuts in the child node.

Even though solver values might indicate better quality we know that a
child nodes quality can not be better than its parent.

Before encoding a pregnant node to send, we set the status of it to
candidate and clear the stored branch object. We do this hoping that
creating branch objects are cheper than sending them through network.