Termination criteria improved. Now we chech for the absolute and relative optimality gap. These should also be reported when the search terminates.
Some of the messages are improved, now they indicate the processor
printing them in "[proc number]" form. "[proc number]" appears
after the debug code. All messages should report their proc number
in this form. In serial mode proc number is 0.

When the subproblem quality is close to incumbent (a fixed number for
now, this will get parametrizied soon), we try to fathom the by
adding cuts. This reduces the number of nodes, cpu time, significantly
in some of the problems. This (deciding when to branch/cut) has more
room for improvement.